Revert "REFACTOR test classes to 'Feature' directory"
This reverts commit 18c886d0
