+++ tests/Unit/AutocompletePathTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Sfneal\Helpers\Aws\S3\Tests\Feature;
+namespace Sfneal\Helpers\Aws\S3\Tests\Unit;
 
 use Illuminate\Support\Facades\Storage;
 use Sfneal\Helpers\Aws\S3\StorageS3;
