+++ tests/Unit/HelpersTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Sfneal\Helpers\Aws\S3\Tests\Feature;
+namespace Sfneal\Helpers\Aws\S3\Tests\Unit;
 
 use GuzzleHttp\Client;
 use Sfneal\Helpers\Aws\S3\StorageS3;
